Ecommerce_app/requirements.txt
Django==5.2.6
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
djangorestframework==3.16.1
pillow==11.3.0
